requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
requests
passlib[bcrypt] 
python-jose
streamlit
python-multipart